LayoutTests/imported/w3c/web-platform-tests/html/browsers/origin/api/origin-from-string.any-expected.txt


PASS Origin.from("") throws a TypeError.
PASS Origin.from("not-valid") throws a TypeError.
PASS Origin.from("about:blank") is an opaque origin.
PASS Origin.from("data:text/plain,opaque") is an opaque origin.
PASS Origin.from("weird-protocol:whatever") is an opaque origin.
PASS Origin.from("weird-hierarchical-protocol://host/path?etc") is an opaque origin.
PASS Origin.from("blob:weird-protocol:whatever") is an opaque origin.
PASS Origin.from("blob:weird-hierarchical-protocol://host/path?etc") is an opaque origin.
PASS Origin.from("http://site.example") is a tuple origin.
PASS Origin.from("https://site.example") is a tuple origin.
PASS Origin.from("https://site.example:123") is a tuple origin.
PASS Origin.from("http://sub.site.example") is a tuple origin.
PASS Origin.from("https://sub.site.example") is a tuple origin.
PASS Origin.from("https://sub.site.example:123") is a tuple origin.
PASS Origin.from("https://xn--mlauted-m2a.example") is a tuple origin.
PASS Origin.from("ftp://ftp.example") is a tuple origin.
PASS Origin.from("ws://ws.example") is a tuple origin.
PASS Origin.from("wss://wss.example") is a tuple origin.
PASS Origin.from("https://trailing.slash/") is a tuple origin.
PASS Origin.from("https://user:pass@site.example") is a tuple origin.
PASS Origin.from("https://has.a.port:1234/and/path") is a tuple origin.
PASS Origin.from("https://ümlauted.example") is a tuple origin.
PASS Origin.from("blob:https://example.com/some-guid") is a tuple origin.
PASS Origin.from("ftp://example.com/") is a tuple origin.
PASS Origin.from("https://example.com/path?query#fragment") is a tuple origin.
PASS Origin.from("https://127.0.0.1/") is a tuple origin.
PASS Origin.from("https://[::1]/") is a tuple origin.
